+++ http4k-security-oauth/src/main/kotlin/org/http4k/security/oauth/server/ResponseRender.kt
@@ -21,7 +21,7 @@ interface ResponseRender {
 
     companion object {
         fun forAuthRequest(authorizationRequest: AuthRequest) =
-            forAuthRequest(authorizationRequest.responseMode, authorizationRequest.responseType, authorizationRequest.redirectUri)
+            forAuthRequest(authorizationRequest.responseMode, authorizationRequest.responseType, authorizationRequest.redirectUri!!)
 
         fun forAuthRequest(responseMode: ResponseMode?, responseType: ResponseType, redirectUri: Uri) =
             when (responseMode) {