+++ http4k-security-oauth/src/main/kotlin/org/http4k/security/oauth/server/request/RequestJWTValidator.kt
@@ -5,12 +5,12 @@ import org.http4k.security.oauth.server.InvalidAuthorizationRequest
 import org.http4k.security.openid.RequestJwtContainer
 import java.lang.UnsupportedOperationException
 
-interface RequestValidator {
+interface RequestJWTValidator {
 
     fun validate(clientId: ClientId, requestJwtContainer: RequestJwtContainer): InvalidAuthorizationRequest?
 
     companion object {
-        val Unsupported = object : RequestValidator {
+        val Unsupported = object : RequestJWTValidator {
             override fun validate(clientId: ClientId, requestJwtContainer: RequestJwtContainer): InvalidAuthorizationRequest? {
                 throw UnsupportedOperationException("Request JWTs are not supported by this server")
             }