+++ http4k-security-oauth/src/main/kotlin/org/http4k/security/oauth/server/request/RequestObject.kt
@@ -18,7 +18,10 @@ data class RequestObject(@JsonProperty("client_id") val client: ClientId? = null
                          @JsonProperty("state") val state: State? = null,
                          @JsonProperty("nonce") val nonce: Nonce? = null,
                          @JsonProperty("max_age") val magAge: Long? = null,
-                         @JsonProperty("claims") val claims: Claims = Claims())
+                         @JsonProperty("claims") val claims: Claims = Claims()) {
+
+    fun scopes(): List<String> = this.scope?.let { it.split(" ") } ?: emptyList()
+}
 
 data class Claims(@JsonProperty("userinfo") val userInfo: Map<String, Claim>? = null,
                   @JsonProperty("id_token") val idToken: Map<String, Claim>? = null)