+++ http4k-security-oauth/src/test/kotlin/org/http4k/security/oauth/server/AuthoriseRequestErrorRenderTest.kt
@@ -16,7 +16,7 @@ import org.http4k.lens.Header.CONTENT_TYPE
 import org.http4k.security.ResponseMode.Query
 import org.http4k.security.ResponseType.Code
 import org.http4k.security.ResponseType.CodeIdToken
-import org.http4k.security.oauth.server.request.RequestValidator
+import org.http4k.security.oauth.server.request.RequestJWTValidator
 import org.http4k.security.openid.RequestJwtContainer
 import org.junit.jupiter.api.Test
 import java.util.Base64
@@ -209,7 +209,7 @@ internal class AuthoriseRequestErrorRenderTest {
 
     }
 
-    private val requestValidator = object : RequestValidator {
+    private val requestValidator = object : RequestJWTValidator {
         override fun validate(clientId: ClientId, requestJwtContainer: RequestJwtContainer): InvalidAuthorizationRequest? {
             return if (requestJwtContainer.value == "inValidRequest") {
                 InvalidAuthorizationRequest("request not correctly signed")