+++ http4k-security-oauth/src/test/kotlin/org/http4k/security/oauth/server/ClientValidationFilterTest.kt
@@ -18,7 +18,7 @@ import org.http4k.format.Jackson
 import org.http4k.hamkrest.hasBody
 import org.http4k.hamkrest.hasStatus
 import org.http4k.security.ResponseType.Code
-import org.http4k.security.oauth.server.request.RequestValidator
+import org.http4k.security.oauth.server.request.RequestJWTValidator
 import org.http4k.security.openid.RequestJwtContainer
 import org.junit.jupiter.api.Test
 
@@ -46,7 +46,7 @@ internal class ClientValidationFilterTest {
 
     }
 
-    private val requestValidator = object : RequestValidator {
+    private val requestValidator = object : RequestJWTValidator {
         override fun validate(clientId: ClientId, requestJwtContainer: RequestJwtContainer): InvalidAuthorizationRequest? {
             return if (requestJwtContainer.value == "inValidRequest") {
                 InvalidAuthorizationRequest("request not correctly signed")